backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
supabase
pypdf
fastembed
faiss-cpu
numpy
groq